+++ express/scripts/utils.js
@@ -105,6 +105,7 @@ export const [setConfig, updateConfig, getConfig] = (() => {
           || 'ltr';
         document.documentElement.setAttribute('dir', dir);
       } catch (e) {
+        // eslint-disable-next-line no-console
         console.log('Invalid or missing locale:', e);
       }
       config.locale.contentRoot = `${origin}${config.locale.prefix}${config.contentRoot ?? ''}`;
@@ -1051,7 +1052,6 @@ function resolveFragments() {
       const $marker = Array.from(document.querySelectorAll('main > div h3'))
         .find(($title) => $title.textContent.trim().toLocaleLowerCase() === marker);
       if (!$marker) {
-        console.log(`no fragment with marker "${marker}" found`);
         return;
       }
       let $fragment = $marker.closest('main > div');
@@ -1063,7 +1063,6 @@ function resolveFragments() {
         $emptyFragment.remove();
       }
       if (!$fragment) {
-        console.log(`no content found for fragment "${marker}"`);
         return;
       }
       setTimeout(() => {
@@ -1071,7 +1070,6 @@ function resolveFragments() {
         Array.from($fragment.children).forEach(($elem) => $cell.appendChild($elem));
         $marker.remove();
         $fragment.remove();
-        console.log(`fragment "${marker}" resolved`);
       }, 500);
     });
 }
@@ -1455,10 +1453,10 @@ export function toCamelCase(name) {
  */
 export function getExperiment() {
   let experiment = toClassName(getMetadata('experiment'));
-
-  if (!/adobe\.com/.test(window.location.hostname) && !/\.hlx\.live/.test(window.location.hostname)) {
+  const { hostname } = window.location;
+  if (!(/adobe\.com/.test(hostname) || /\.hlx\.live/.test(hostname) || hostname.includes('localhost'))) {
     experiment = '';
-    // reason = 'not prod host';
+    // reason = 'not prod host and not local';
   }
   if (window.location.hash) {
     experiment = '';
@@ -1567,62 +1565,45 @@ export async function getExperimentConfig(experimentId) {
       });
       config.variants = variants;
       config.variantNames = variantNames;
-      console.log(config);
       return config;
     } catch (e) {
+      // eslint-disable-next-line no-console
       console.log('error loading experiment manifest: %s', path, e);
     }
     return null;
   }
 }
 
-/**
- * Replaces element with content from path
- * @param {string} path
- * @param {HTMLElement} element
- */
-async function replaceInner(path, element) {
-  const plainPath = `${path}.plain.html`;
-  try {
-    const resp = await fetch(plainPath);
-    const html = await resp.text();
-    element.innerHTML = html;
-  } catch (e) {
-    console.log(`error loading experiment content: ${plainPath}`, e);
+async function loadAndRunExp(config, forcedExperiment, forcedVariant) {
+  const promises = [import('./experiment.js')];
+  if (getMetadata('aepaudience') === 'on') {
+    // only aep exps: rush alloy thru launch
+    promises.push(loadScript('/express/scripts/instrument.js', 'module'));
+    let alloyLoadingResolver;
+    window.alloyLoader = new Promise((resolve) => {
+      alloyLoadingResolver = resolve;
+    });
+    window.addEventListener('alloy_sendEvent', (e) => {
+      // fired by launch loaded by martech loaded by instrument
+      if (e.detail.type === 'pageView') {
+        window.alloyLoaded = true;
+        alloyLoadingResolver(e.detail.result);
+      }
+    });
+    // tolerate max 5s for exp overheads
+    setTimeout(() => {
+      if (!window.alloyLoaded) {
+        // window.lana.log('Alloy failed to load', {
+        //   errorType: 'e',
+        // });
+        // eslint-disable-next-line no-console
+        console.error('Alloy failed to load');
+        alloyLoadingResolver();
+      }
+    }, 5000);
   }
-  return null;
-}
-
-/**
- * Generates a decision policy object which is understood by UED from an
- * experiment configuration.
- * @param {*} config Experiment configuration
- * @returns Experiment decision policy object to be passed to UED.
- */
-function getDecisionPolicy(config) {
-  const decisionPolicy = {
-    id: 'content-experimentation-policy',
-    rootDecisionNodeId: 'n1',
-    decisionNodes: [{
-      id: 'n1',
-      type: 'EXPERIMENTATION',
-      experiment: {
-        id: config.id,
-        identityNamespace: 'ECID',
-        randomizationUnit: 'DEVICE',
-        treatments: Object.entries(config.variants).map(([key, props]) => ({
-          id: key,
-          allocationPercentage: props.percentageSplit
-            ? parseFloat(props.percentageSplit) * 100
-            : 100 - Object.values(config.variants).reduce((result, variant) => {
-              const returnResult = result - (parseFloat(variant.percentageSplit || 0) * 100);
-              return returnResult;
-            }, 100),
-        })),
-      },
-    }],
-  };
-  return decisionPolicy;
+  const [{ runExps }] = await Promise.all(promises);
+  await runExps(config, forcedExperiment, forcedVariant);
 }
 
 /**
@@ -1630,7 +1611,6 @@ function getDecisionPolicy(config) {
  */
 async function decorateTesting() {
   try {
-    // let reason = '';
     const usp = new URLSearchParams(window.location.search);
 
     const experiment = getExperiment();
@@ -1639,68 +1619,7 @@ async function decorateTesting() {
     if (experiment) {
       const config = await getExperimentConfig(experiment);
       if (config && (toCamelCase(config.status) === 'active' || forcedExperiment)) {
-        window.addEventListener('alloy_sendEvent', (e) => {
-          if (e.detail.type === 'pageView') {
-            console.log(e.detail);
-            console.log(e.detail.result);
-          }
-        });
-        // rush launch for alloy configuration
-        await loadScript('/express/scripts/instrument.js', 'module');
-        const { DEFAULT_EXPERIMENT_OPTIONS, AUDIENCES, getResolvedAudiences } = await import('./experiment.js');
-        const experimentOptions = {
-          ...DEFAULT_EXPERIMENT_OPTIONS,
-          ...{ audiences: AUDIENCES },
-        };
-        config.resolvedAudiences = await getResolvedAudiences(config.audience.split(',').map((a) => a.trim()), experimentOptions);
-        config.run = forcedExperiment
-          || !config.resolvedAudiences
-          || config.resolvedAudiences.length;
-
-        window.hlx = window.hlx || {};
-        if (config.run) {
-          window.hlx.experiment = config;
-          if (forcedVariant && config.variantNames.includes(forcedVariant)) {
-            config.selectedVariant = forcedVariant;
-          } else {
-            const ued = await import('./ued/ued-0.2.0.js');
-            const decision = ued.evaluateDecisionPolicy(getDecisionPolicy(config), {});
-            config.selectedVariant = decision.items[0].id;
-          }
-          sampleRUM('experiment', { source: config.id, target: config.selectedVariant });
-          // populate ttMETA with hlx experimentation details
-          window.ttMETA = window.ttMETA || [];
-          const experimentDetails = {
-            CampaignId: window.hlx.experiment.id,
-            CampaignName: window.hlx.experiment.experimentName,
-            OfferId: window.hlx.experiment.selectedVariant,
-            OfferName: window.hlx.experiment.variants[window.hlx.experiment.selectedVariant].label,
-          };
-          window.ttMETA.push(experimentDetails);
-          // add hlx experiment details as dynamic variables
-          // for Content Square integration
-          // eslint-disable-next-line no-underscore-dangle
-          if (window._uxa) {
-            for (const propName of Object.keys(experimentDetails)) {
-              // eslint-disable-next-line no-underscore-dangle
-              window._uxa.push(['trackDynamicVariable', { key: propName, value: experimentDetails[propName] }]);
-            }
-          }
-          if (config.selectedVariant !== 'control') {
-            const currentPath = window.location.pathname;
-            const pageIndex = config.variants.control.pages.indexOf(currentPath);
-            if (pageIndex >= 0) {
-              const page = config.variants[config.selectedVariant].pages[pageIndex];
-              if (page) {
-                const experimentPath = new URL(page, window.location.href).pathname.split('.')[0];
-                if (experimentPath && experimentPath !== currentPath) {
-                  await replaceInner(experimentPath, document.querySelector('main'));
-                  removeIrrelevantSections(document.querySelector('main'));
-                }
-              }
-            }
-          }
-        }
+        await loadAndRunExp(config, forcedExperiment, forcedVariant);
       }
     }
     const martech = usp.get('martech');
@@ -1710,6 +1629,7 @@ async function decorateTesting() {
       loadScript('/express/scripts/instrument.js', 'module');
     }
   } catch (e) {
+    // eslint-disable-next-line no-console
     console.log('error testing', e);
   }
 }
@@ -2127,6 +2047,7 @@ function displayEnv() {
         setHelixEnv('stage', { spark: url.host });
       }
       if (window.location.hostname !== url.hostname) {
+        // eslint-disable-next-line no-console
         console.log(`external referrer detected: ${document.referrer}`);
       }
     }
@@ -2138,6 +2059,7 @@ function displayEnv() {
       document.body.appendChild($helixEnv);
     }
   } catch (e) {
+    // eslint-disable-next-line no-console
     console.log(`display env failed: ${e.message}`);
   }
 }
